Revert "Consolidate celo changes in one file"
This reverts commit a37e58f286988b325b3e0d12aaea206dd6081e84.
